requirements.txt
# Core
llama-index==0.10.38
llama-index-llms-openai==0.1.19
llama-index-embeddings-openai==0.1.10
llama-index-vector-stores-chroma==0.1.9

# OpenAI & HTTP client
openai==1.30.1
httpx==0.27.2

# Vector DB
chromadb==0.5.4

# UI & data
streamlit==1.33.0
pandas==2.1.4
python-dotenv==1.0.0
